+++ utoipa-gen/src/component/features/attributes.rs
@@ -258,12 +258,12 @@ impl From<ReadOnly> for Feature {
 impl_feature! {
     #[derive(Clone)]
     #[cfg_attr(feature = "debug", derive(Debug))]
-    pub struct Title(String);
+    pub struct Title(parse_utils::LitStrOrExpr);
 }
 
 impl Parse for Title {
     fn parse(input: syn::parse::ParseStream, _: Ident) -> syn::Result<Self> {
-        parse_utils::parse_next_literal_str(input).map(Self)
+        parse_utils::parse_next_literal_str_or_expr(input).map(Self)
     }
 }
 
@@ -542,11 +542,11 @@ impl_feature! {"names" =>
     /// Specify names of unnamed fields with `names(...) attribute for `IntoParams` derive.
     #[cfg_attr(feature = "debug", derive(Debug))]
     #[derive(Clone)]
-    pub struct IntoParamsNames(Vec<String>);
+    pub struct IntoParamsNames(Vec<parse_utils::LitStrOrExpr>);
 }
 
 impl IntoParamsNames {
-    pub fn into_values(self) -> Vec<String> {
+    pub fn into_values(self) -> Vec<parse_utils::LitStrOrExpr> {
         self.0
     }
 }
@@ -554,9 +554,9 @@ impl IntoParamsNames {
 impl Parse for IntoParamsNames {
     fn parse(input: syn::parse::ParseStream, _: Ident) -> syn::Result<Self> {
         Ok(Self(
-            parse_utils::parse_comma_separated_within_parenthesis::<LitStr>(input)?
-                .iter()
-                .map(LitStr::value)
+            parse_utils::parse_comma_separated_within_parenthesis::<LitStrOrExpr>(input)?
+                .into_iter()
+                .map(|lit| parse_utils::LitStrOrExpr::from(lit))
                 .collect(),
         ))
     }