+++ utoipa-gen/src/component/into_params.rs
@@ -25,6 +25,7 @@ use crate::{
         FieldRename,
     },
     doc_comment::CommentAttributes,
+    parse_utils::LitStrOrExpr,
     token_stream::{quote_diagnostics, Diagnostics, ToTokensDiagnostics},
     OptionExt, Required,
 };
@@ -204,7 +205,7 @@ fn parse_field_features(field: &Field) -> Result<Vec<Feature>, Diagnostics> {
 impl IntoParams {
     fn get_struct_fields(
         &self,
-        field_names: &Option<&Vec<String>>,
+        field_names: &Option<&Vec<LitStrOrExpr>>,
     ) -> Result<impl Iterator<Item = &Field>, Diagnostics> {
         let ident = &self.ident;
         match &self.data {
@@ -239,7 +240,7 @@ impl IntoParams {
     fn validate_unnamed_field_names(
         &self,
         unnamed_fields: &Punctuated<Field, Comma>,
-        field_names: &Option<&Vec<String>>,
+        field_names: &Option<&Vec<LitStrOrExpr>>,
     ) -> Option<Diagnostics> {
         let ident = &self.ident;
         match field_names {
@@ -278,7 +279,7 @@ pub struct FieldParamContainerAttributes<'a> {
     /// See [`IntoParamsAttr::style`].
     style: &'a Option<Feature>,
     /// See [`IntoParamsAttr::names`]. The name that applies to this field.
-    name: Option<&'a String>,
+    name: Option<&'a LitStrOrExpr>,
     /// See [`IntoParamsAttr::parameter_in`].
     parameter_in: &'a Option<Feature>,
     /// Custom rename all if serde attribute is not present.
@@ -342,18 +343,14 @@ impl Param {
         let mut tokens = TokenStream::new();
         let field_serde_params = &field_serde_params;
         let ident = &field.ident;
-        let mut name = &*ident
+        let mut name = ident
             .as_ref()
-            .map(|ident| ident.to_string())
+            .map(|ident| LitStrOrExpr::from(ident.to_string()))
             .or_else(|| container_attributes.name.cloned())
             .ok_or_else(||
                 Diagnostics::with_span(field.span(), "No name specified for unnamed field.")
                     .help("Try adding #[into_params(names(...))] container attribute to specify the name for this field")
             )?;
-
-        if name.starts_with("r#") {
-            name = &name[2..];
-        }
 
         let (schema_features, mut param_features) =
             Param::resolve_field_features(field_features, &container_attributes)
@@ -376,8 +373,22 @@ impl Param {
         let rename_all = serde_container.rename_all.as_ref().or(container_attributes
             .rename_all
             .map(|rename_all| rename_all.as_rename_rule()));
-        let name = super::rename::<FieldRename>(name, rename_to, rename_all)
-            .unwrap_or(Cow::Borrowed(name));
+
+        if let LitStrOrExpr::LitStr(s) = &name {
+            let lit_value = s.value();
+            let value = if lit_value.starts_with("r#") {
+                Cow::Borrowed(&lit_value[2..])
+            } else {
+                Cow::Borrowed(lit_value.as_str())
+            };
+
+            name = LitStrOrExpr::from(
+                super::rename::<FieldRename>(&value, rename_to, rename_all)
+                    .unwrap_or(value)
+                    .into_owned(),
+            );
+        }
+
         let type_tree = TypeTree::from_type(&field.ty)?;
 
         tokens.extend(quote! { utoipa::openapi::path::ParameterBuilder::new()